requirements.txt
asgiref==3.8.1
Django==4.2.14
django-environ==0.11.2
django-guardian==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
